Fixing the use case "Manage Projects -> 3. Manage a project as a PI"
The original acceptance tests stated:
"Some other roles should be set for other general users
who are only going to use the allocated NERC resources."
We discussed as a team what other roles where suggested
by Leo McGann who is no longer on the project.

The roles in Coldfront include User and Manager,
but from the OpenShift side, Users and Managers will have the same access of "edit" on the project.
We decided that this wording can be removed.